Split some stuff up into it's own file
